sender/requirements.txt
python-dotenv
redis
Flask
Flask-Session
gunicorn